requirements.txt
# Core dependencies
nodriver
google-genai
openai
python-dotenv
requests
beautifulsoup4
filelock>=3.12.0
